added new file, added models, and updated other files
I added a predict file that is able to load and test any of the saved models against acutal subjects in the excel sheet. You give it the model (like you would the preNeuralNetwork.py), and the number of people to predict for, and it will spit out results.
I clearned up preNeuralNetwork.py. I changed debug from being boolean to now being a degree of debug(int) with 1 being default. I preffer 2.
I added a learning rate that you can now give as input to use.
I added the ability to normalize data. I need to test this though to see if it helps later on as it looked liked it hurt it when i initaly tested it.
Changed rmsprop to Adam.
I changed preprocessor to take chunks of the data out for MIC category 0, 1, and 3 of 90%-85% so the data can be even. This takes our amount of data from 103,000 to about 14,000, but it does not just predict 0 and 1 now. It sometimes will predict 2.