Tuned network, and applied changes to files
Made it simpler to tune by adding a for loop for drop out and a for loop for learning rate.
Fixed predict and preNeuralNetwork so that they can work on a GPU.
Fixed preNeuralNetwork so that it can properly do sequential trains on different models.
Added EarlyStopping to stop the networks when they plateu automatically.
Added CSV logger so that it can log all metrics for each model to a csv file to be manually inspected later.
Added a Batch Normalization to the input layer (Have not tried it yet)
Added a batch file for my desktop to be able to more easily start up tensorboard.