fixed a bug and fixed batch norm
I made the 1 into a 1.0 for the drop out option. This will allow the models to be loaded in normally.
Fixed batch normalization so that it is now on every layer between linear and activation layer.
Fixed a bug in the model load that it did not look for recall and precision.